modified client to send in public_decrypt
