+++ encryption_lib.c
@@ -33,14 +33,63 @@ RSA *RSA_from_file(char* file_path, int pub_priv) {
 //}
 
 
-int public_encrypt(unsigned char* data, int data_len, unsigned char* file_path, unsigned char* encrypted) {
+int public_encrypt(unsigned char* data, int data_len, unsigned char* file_path,
+    unsigned char* encrypted, int sockfd) {
 	RSA *rsa = RSA_from_file(file_path, 1);		//file name?
+    
+    // Allocate the proper size for our ctext
+    encrypted = (unsigned char *) malloc(RSA_size(rsa));
 	int result = RSA_public_encrypt(data_len, data, encrypted, rsa, RSA_PKCS1_OAEP_PADDING);
-	return result;
+    
+    // moved send into here to actually send a ctext
+    send(sockfd, (const void *) encrypted, result, 0);
+    
+    if (result == -1) {
+        printf("%s\n", "Error RSA encrypt pub");
+        char err[130];
+        ERR_load_crypto_strings();
+        ERR_error_string(ERR_get_error(), err);
+        fprintf(stderr, "Error encrypting message: %s\n", err);
+    }
+
+	// moved send into here to actually send a ctext
+    send(sockfd, (const void *) encrypted, result, 0);
+    
+    return result;
 }
 
 int private_decrypt(unsigned char* data, int data_len, unsigned char* file_path, unsigned char* decrypted) {
 	RSA *rsa = RSA_from_file(file_path, 0);		//file name?
-	int result = RSA_private_decrypt(data_len, data, decrypted, rsa, RSA_PKCS1_OAEP_PADDING);
+	int result = RSA_private_decrypt(data_len, data, decrypted, rsa, RSA_PKCS1_PADDING);
 	return result;
-}	
+}
+
+// Old method of delivering a message to stdout
+void *deliver_ctext(void *arg) {
+    FILE* ctext_fd;
+    int i = 0, j = 0;
+    struct stat file_info;
+    long last_modifed_time = -1;
+    char *buff = NULL;
+    size_t line_buf_size = 0;
+    ssize_t line_size;
+
+    // Find the most recent ctext in the log file
+    while (1) {
+        ctext_fd = fopen("log.txt", "r");
+        line_size = 1;
+        j = 1;
+        while (line_size >= 0) {
+            line_size = getline(&buff, &line_buf_size, ctext_fd);
+            j++;
+            if (j > i) {
+                printf("%s   ", buff);
+            }
+        }
+        // Update the i (current line) value and close the file
+        i = j;
+        fclose(ctext_fd);
+    }
+
+}
+
