+++ encryption_lib.h
@@ -17,7 +17,7 @@ struct Message{
 RSA *RSA_from_file(char* file_path, int pub_priv);
 //int RSA_public_encrypt(int length, unsigned char* ptxt, unsigned char* ctxt, RSA *rsa, int pad);
 //int RSA_private_decrypt(int length, unsigned char* ctxt, unsigned char* ptxt, RSA *rsa, int pad);
-int public_encrypt(unsigned char* data, int data_len, unsigned char* key, unsigned char* encrypted);
+int public_encrypt(unsigned char* data, int data_len, unsigned char* key, unsigned char* encrypted, int sockfd);
 int private_decrypt(unsigned char* enc_data, int data_len, unsigned char* key, unsigned char* decrypted);
 
 #endif
